check that only .key column exists instead of just counting columns
this catches the same error it just removes the need for a code comment. that is, writing the check this way lets us remove the comment "#Mapped data is created with just a .key column in the SurveyData obj" because the error check is written in a way that's makes it clear that '.key' is the only column if mapping() hasn't been called.