Minor tweaks throughout
Hi Dewi,

First, I just wanted to say how blown away I was by how great this was. I can tell how much work you've put into this, and it has paid off.

In general, the changes that I made were just small things around language use. My general approach is to try to remove as many words as possible. For instance, sometimes you'd say 'Note...'. Usually that 'Note' part can be removed without any lost meaning. Similarly, there were cases where three or four words could be replaced with one or even removed. For instance, 'it is the case', can usually just be removed.

Another tweak that I made was to try to make things a little more explicit. It's completely normal, when you've spent so much time on something, that there are aspects that are self-explanatory to you, that are less so for someone coming to this for the first time.

With the code - I just auto-formatted it all with CMD+SHIFT+A. So there are various line breaks and things that have been added, but there's no actual code edits.

While it may look like a lot of edits, it's just a bunch of tiny changes. Thank you so much for this great vignette. I found it really useful and I know that our users will to.

Rohan